requirements.txt
yt-dlp
yt-dlp-ejs
biliup
PyYAML
flask
Pillow
requests
